pyproject.toml
[tool.poetry]
name = "crawler-datajud"
version = "0.1.0"
description = ""
authors = ["v3ntur4X <51869757+v3ntur4X@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.scripts]
crawler-datajud = "crawler_datajud.cli:app"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
typer = "^0.15.2"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
